+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Added `cfg logout` command, which iterates through fields in your config file which contain sensitive data and removes their value
 - Added `cfg location` command which prints config file path.
 - Added `cfg whoami` command which prints the authenticated user, with the option (`-p/--provenance`) to print the source and rules/order  of credentials.
-
+- Added `--ftp-user`, `--ftp-pass` to image upload command, which doesn't support JWT token. They can be used to override auth credentials only for the FTP server.
 
 ### Changed
 - `ionosctl login` is now deprecated: it will simply call `ionosctl cfg login`.